backend/analytics/technical/patterns/elliott_wave.py
"""
Elliott Wave Theory Implementation
Advanced wave pattern recognition for market analysis
"""

from typing import List, Dict, Optional, Tuple
import pandas as pd
import numpy as np
from dataclasses import dataclass
from enum import Enum
import scipy.signal as signal


class WaveType(Enum):
    """Types of Elliott Waves"""
    IMPULSE = "impulse"
    CORRECTIVE = "corrective"
    DIAGONAL = "diagonal"
    TRIANGLE = "triangle"
    FLAT = "flat"
    ZIGZAG = "zigzag"


@dataclass
class ElliottWave:
    """Represents an Elliott Wave pattern"""
    wave_type: WaveType
    degree: str  # Grand, Super, Cycle, Primary, etc.
    wave_number: int  # 1-5 for impulse, A-B-C for corrective
    start_point: Tuple[int, float]
    end_point: Tuple[int, float]
    subwaves: List['ElliottWave']
    confidence: float


class ElliottWaveAnalyzer:
    """Analyzes price data for Elliott Wave patterns"""
    
    def __init__(self):
        self.fibonacci_ratios = [0.236, 0.382, 0.5, 0.618, 0.786, 1.0, 1.272, 1.618, 2.618]
        self.wave_rules = {
            'impulse': self._validate_impulse_wave,
            'corrective': self._validate_corrective_wave,
            'diagonal': self._validate_diagonal_wave
        }
    
    def identify_waves(self, df: pd.DataFrame) -> List[ElliottWave]:
        """Identify Elliott Wave patterns in price data"""
        # Find significant peaks and troughs
        peaks, troughs = self._find_turning_points(df)
        
        # Identify potential wave patterns
        waves = []
        
        # Check for impulse waves
        impulse_waves = self._identify_impulse_waves(df, peaks, troughs)
        waves.extend(impulse_waves)
        
        # Check for corrective waves
        corrective_waves = self._identify_corrective_waves(df, peaks, troughs)
        waves.extend(corrective_waves)
        
        return waves
    
    def _find_turning_points(self, df: pd.DataFrame, prominence: float = 0.05) -> Tuple[np.array, np.array]:
        """Find significant peaks and troughs in price data"""
        prices = df['close'].values
        
        # Find peaks
        peaks, peak_props = signal.find_peaks(prices, prominence=prices.mean() * prominence)
        
        # Find troughs (invert and find peaks)
        troughs, trough_props = signal.find_peaks(-prices, prominence=prices.mean() * prominence)
        
        return peaks, troughs
    
    def _identify_impulse_waves(self, df: pd.DataFrame, peaks: np.array, troughs: np.array) -> List[ElliottWave]:
        """Identify 5-wave impulse patterns"""
        waves = []
        
        # Need at least 6 turning points for a 5-wave pattern
        turning_points = sorted(list(peaks) + list(troughs))
        
        if len(turning_points) < 6:
            return waves
        
        for i in range(len(turning_points) - 5):
            segment = turning_points[i:i+6]
            prices = [df['close'].iloc[idx] for idx in segment]
            
            # Check if this could be an impulse wave
            if self._is_impulse_pattern(prices):
                wave = ElliottWave(
                    wave_type=WaveType.IMPULSE,
                    degree=self._determine_degree(df, segment[0], segment[-1]),
                    wave_number=5,
                    start_point=(segment[0], prices[0]),
                    end_point=(segment[-1], prices[-1]),
                    subwaves=self._create_subwaves(segment, prices),
                    confidence=self._calculate_impulse_confidence(prices)
                )
                waves.append(wave)
        
        return waves
    
    def _identify_corrective_waves(self, df: pd.DataFrame, peaks: np.array, troughs: np.array) -> List[ElliottWave]:
        """Identify A-B-C corrective patterns"""
        waves = []
        
        # Need at least 4 turning points for an A-B-C pattern
        turning_points = sorted(list(peaks) + list(troughs))
        
        if len(turning_points) < 4:
            return waves
        
        for i in range(len(turning_points) - 3):
            segment = turning_points[i:i+4]
            prices = [df['close'].iloc[idx] for idx in segment]
            
            # Check if this could be a corrective wave
            if self._is_corrective_pattern(prices):
                wave = ElliottWave(
                    wave_type=WaveType.CORRECTIVE,
                    degree=self._determine_degree(df, segment[0], segment[-1]),
                    wave_number=3,  # A-B-C
                    start_point=(segment[0], prices[0]),
                    end_point=(segment[-1], prices[-1]),
                    subwaves=self._create_corrective_subwaves(segment, prices),
                    confidence=self._calculate_corrective_confidence(prices)
                )
                waves.append(wave)
        
        return waves
    
    def _is_impulse_pattern(self, prices: List[float]) -> bool:
        """Check if price pattern follows impulse wave rules"""
        if len(prices) != 6:
            return False
        
        # Wave 2 cannot retrace more than 100% of Wave 1
        wave2_retrace = abs(prices[2] - prices[1]) / abs(prices[1] - prices[0])
        if wave2_retrace >= 1.0:
            return False
        
        # Wave 3 cannot be the shortest
        wave1_length = abs(prices[1] - prices[0])
        wave3_length = abs(prices[3] - prices[2])
        wave5_length = abs(prices[5] - prices[4])
        
        if wave3_length < wave1_length and wave3_length < wave5_length:
            return False
        
        # Wave 4 cannot overlap Wave 1 price territory
        if prices[0] < prices[1]:  # Bullish impulse
            if prices[4] < prices[1]:
                return False
        else:  # Bearish impulse
            if prices[4] > prices[1]:
                return False
        
        return True
    
    def _is_corrective_pattern(self, prices: List[float]) -> bool:
        """Check if price pattern follows corrective wave rules"""
        if len(prices) != 4:
            return False
        
        # Basic A-B-C pattern validation
        # Wave B typically retraces 38.2% to 138.2% of Wave A
        wave_a = prices[1] - prices[0]
        wave_b = prices[2] - prices[1]
        
        if wave_a != 0:
            b_retrace = abs(wave_b / wave_a)
            if b_retrace < 0.236 or b_retrace > 1.618:
                return False
        
        return True
    
    def _determine_degree(self, df: pd.DataFrame, start_idx: int, end_idx: int) -> str:
        """Determine the degree of the wave based on timeframe"""
        duration = end_idx - start_idx
        
        if duration > 200:
            return "Grand Super Cycle"
        elif duration > 100:
            return "Super Cycle"
        elif duration > 50:
            return "Cycle"
        elif duration > 20:
            return "Primary"
        elif duration > 10:
            return "Intermediate"
        elif duration > 5:
            return "Minor"
        else:
            return "Minute"
    
    def _create_subwaves(self, indices: List[int], prices: List[float]) -> List[ElliottWave]:
        """Create subwave objects for impulse wave"""
        subwaves = []
        
        for i in range(len(indices) - 1):
            subwave = ElliottWave(
                wave_type=WaveType.IMPULSE,
                degree="Subwave",
                wave_number=i + 1,
                start_point=(indices[i], prices[i]),
                end_point=(indices[i+1], prices[i+1]),
                subwaves=[],
                confidence=0.7
            )
            subwaves.append(subwave)
        
        return subwaves
    
    def _create_corrective_subwaves(self, indices: List[int], prices: List[float]) -> List[ElliottWave]:
        """Create subwave objects for corrective wave"""
        subwaves = []
        wave_labels = ['A', 'B', 'C']
        
        for i in range(len(indices) - 1):
            subwave = ElliottWave(
                wave_type=WaveType.CORRECTIVE,
                degree="Subwave",
                wave_number=i,  # Will map to A, B, C
                start_point=(indices[i], prices[i]),
                end_point=(indices[i+1], prices[i+1]),
                subwaves=[],
                confidence=0.7
            )
            subwaves.append(subwave)
        
        return subwaves
    
    def _calculate_impulse_confidence(self, prices: List[float]) -> float:
        """Calculate confidence score for impulse wave"""
        confidence = 0.5
        
        # Check Fibonacci relationships
        wave1 = abs(prices[1] - prices[0])
        wave3 = abs(prices[3] - prices[2])
        wave5 = abs(prices[5] - prices[4])
        
        # Wave 3 often extends to 1.618 * Wave 1
        if wave1 > 0:
            ratio3 = wave3 / wave1
            if 1.5 < ratio3 < 1.8:
                confidence += 0.2
        
        # Wave 5 often equals Wave 1
        if wave1 > 0:
            ratio5 = wave5 / wave1
            if 0.9 < ratio5 < 1.1:
                confidence += 0.2
        
        return min(confidence, 1.0)
    
    def _calculate_corrective_confidence(self, prices: List[float]) -> float:
        """Calculate confidence score for corrective wave"""
        confidence = 0.5
        
        # Check Fibonacci relationships
        wave_a = abs(prices[1] - prices[0])
        wave_b = abs(prices[2] - prices[1])
        wave_c = abs(prices[3] - prices[2])
        
        # Wave C often equals Wave A
        if wave_a > 0:
            ratio = wave_c / wave_a
            if 0.9 < ratio < 1.1:
                confidence += 0.3
        
        return min(confidence, 1.0)
    
    def _validate_impulse_wave(self, wave: ElliottWave) -> bool:
        """Validate impulse wave against Elliott Wave rules"""
        # Implementation of detailed validation rules
        return True
    
    def _validate_corrective_wave(self, wave: ElliottWave) -> bool:
        """Validate corrective wave against Elliott Wave rules"""
        # Implementation of detailed validation rules
        return True
    
    def _validate_diagonal_wave(self, wave: ElliottWave) -> bool:
        """Validate diagonal wave against Elliott Wave rules"""
        # Implementation of detailed validation rules
        return True
    
    def project_next_wave(self, current_wave: ElliottWave) -> Dict[str, float]:
        """Project the likely targets for the next wave"""
        projections = {}
        
        if current_wave.wave_type == WaveType.IMPULSE:
            # Project based on Fibonacci extensions
            wave_length = abs(current_wave.end_point[1] - current_wave.start_point[1])
            
            projections['conservative'] = current_wave.end_point[1] + wave_length * 0.618
            projections['moderate'] = current_wave.end_point[1] + wave_length * 1.0
            projections['aggressive'] = current_wave.end_point[1] + wave_length * 1.618
        
        return projections
